server/requirements.txt
django>=5.2,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0
django-storages[boto3]>=1.14,<2.0
Pillow>=10.0,<11.0
pyotp>=2.9,<3.0
qrcode>=7.4,<8.0
python-decouple>=3.8,<4.0
smile-id-core>=3.0,<4.0
gunicorn>=22.0,<23.0
psycopg2-binary>=2.9,<3.0
whitenoise>=6.6,<7.0
